RoArmM3v1/requirements.txt
# RoArm M3 Professional Control System
# Python Dependencies

# Core dependencies
pyserial>=3.5
pyyaml>=6.0
numpy>=1.20.0
colorama>=0.4.4

# Optional but recommended
matplotlib>=3.5.0
pandas>=1.3.0

# Development tools (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0